fixed a bug and changed the code in some files

+++ src/component/DeleteCategory/DeleteCategory.js
@@ -1,11 +1,10 @@
-import { deleteDoc, doc } from "firebase/firestore";
 import { useContext } from "react";
 import { AppContext } from "../../App";
+import { deleteDoc, doc } from "firebase/firestore";
 import { db } from "../../firebase";
-import "./DeleteCategory.css";
 
-export default function DeleteCategory ({ category }) {
-  const { user } = useContext(AppContext);
+export default function DeleteCategory({ category }) {
+  const { user, products } = useContext(AppContext);
 
   if (!user || !user.isAdmin) {
     return null;
@@ -19,10 +18,17 @@ export default function DeleteCategory ({ category }) {
 
       return;
     }
+
+    if (!window.confirm("Are you sure you want to delete this category?")) {
+      return;
+    }
+
     deleteDoc(doc(db, "categories", category.id));
   }
 
   return (
-    <button className="DeleteCategory" onClick={onDeleteClick}>-</button>
+    <button className="DeleteCategory" onClick={onDeleteClick}>
+      -
+    </button>
   );
 }
